+++ lib/learn/pronounce_learn/today_learn_word_list.dart
@@ -5,61 +5,67 @@ import 'package:flutter_svg/svg.dart';
 import 'package:get/get.dart';
 import 'package:http/http.dart' as http;
 import 'package:loading_animation_widget/loading_animation_widget.dart';
-import 'package:saera/learn/accent_learn/presentation/accent_todaylearn_screen.dart';
-import 'package:saera/learn/search_learn/presentation/widgets/response_statement.dart';
+import 'package:saera/home/presentation/home_screen.dart';
 import 'package:saera/server.dart';
 import 'package:saera/style/color.dart';
 import 'package:saera/style/font.dart';
 
-import 'home/presentation/home_screen.dart';
-import 'login/data/authentication_manager.dart';
+import 'pronounce_learn_screen.dart';
+import '../../login/data/authentication_manager.dart';
 
-class TodayLearnStatementListPage extends StatefulWidget {
+class TodayLearnWordListPage extends StatefulWidget {
 
-  final List<int> sentenceList;
+  final List<int> wordList;
+  final bool isTodayWord;
 
-  const TodayLearnStatementListPage({Key? key, required this.sentenceList});
+  const TodayLearnWordListPage({Key? key, required this.wordList, required this.isTodayWord});
 
   @override
-  State<StatefulWidget> createState() => _TodayLearnStatementListPageState();
+  State<StatefulWidget> createState() => _TodayLearnWordListPageState();
 }
 
-class _TodayLearnStatementListPageState extends State<TodayLearnStatementListPage> {
+class _TodayLearnWordListPageState extends State<TodayLearnWordListPage> {
   final AuthenticationManager _authManager = Get.find();
 
-  Future<dynamic>? statement;
+  Future<dynamic>? word;
 
-  Future<List<Statement>> getStatement() async {
-    List<Statement> statementList = [];
-    String statement_id = "";
-    for(int i = 0; i<widget.sentenceList.length; i++) {
-      statement_id += 'idList=${widget.sentenceList[i]}&';
+  Color selectTagColor(String tag) { // 태그별 컬러파레트 필요
+    if (tag == '구개음화') {
+      return ColorStyles.saeraBlue.withOpacity(0.5);
+    } else if (tag == "ㄴ첨가") {
+      return ColorStyles.saeraKhaki.withOpacity(0.5);
+    } else if (tag == '두음법칙') {
+      return ColorStyles.saeraPink3.withOpacity(0.5);
+    } else if (tag == '치조마찰음화') {
+      return ColorStyles.saeraYellow.withOpacity(0.5);
+    } else if (tag == '단모음화') {
+      return ColorStyles.saeraOlive1.withOpacity(0.5);
+    } else {
+      return ColorStyles.saeraBeige.withOpacity(0.5);
     }
-    var url = Uri.parse('$serverHttp/complete?type=STATEMENT&$statement_id');
+  }
+
+  Future<List<Word>> getWord() async {
+    List<Word> wordList = [];
+    String word_id = "";
+    for(int i = 0; i<widget.wordList.length; i++) {
+      word_id += 'idList=${widget.wordList[i]}&';
+    }
+    var url = Uri.parse('$serverHttp/complete?type=WORD&$word_id&isTodayStudy=${widget.isTodayWord}');
     print("url : $url");
     final response = await http.get(url, headers: {'accept': 'application/json', "content-type": "application/json", "authorization" : "Bearer ${_authManager.getToken()}"});
     if (response.statusCode == 200) {
       var body = jsonDecode(utf8.decode(response.bodyBytes));
       for (dynamic i in body) {
         int id = i["id"];
-        String content = i["content"];
-        List<String> tags = List.from(i["tags"]);
+        String notation = i["notation"];
+        String pronunciation = i["pronunciation"];
+        String tag = i["tag"];
         bool bookmarked = i["bookmarked"];
-        bool recommended = i["recommended"];
-        statementList.add(Statement(id: id, content: content, tags: tags, bookmarked: bookmarked, recommended: recommended));
+        wordList.add(Word(id: id, notation: notation, pronunciation: pronunciation, tag: tag, bookmarked: bookmarked));
       }
     }
-    return statementList;
-  }
-
-  Color selectTagColor(String tag) {
-    if (tag == '일상' || tag == '소비' || tag == '인사' || tag == '은행/공공기관' || tag == '회사') {
-      return ColorStyles.saeraBlue.withOpacity(0.5);
-    } else if (tag == '의문문' || tag == '존댓말' || tag == '부정문' || tag == '감정 표현') {
-      return ColorStyles.saeraBeige.withOpacity(0.5);
-    } else {
-      return ColorStyles.saeraYellow.withOpacity(0.5);
-    }
+    return wordList;
   }
 
   createBookmark (int id) async {
@@ -76,7 +82,7 @@ class _TodayLearnStatementListPageState extends State<TodayLearnStatementListPag
 
   @override
   void initState() {
-    statement = getStatement();
+    word = getWord();
     super.initState();
   }
 
@@ -106,14 +112,13 @@ class _TodayLearnStatementListPageState extends State<TodayLearnStatementListPag
     Widget todayLearnWordTextSection = Container(
       margin: const EdgeInsets.only(left: 10.0),
       padding: EdgeInsets.only(top: MediaQuery.of(context).size.height*0.03),
-      child: const Text(
-        '오늘 학습한 문장 목록',
-        style: TextStyles.xxLargeTextStyle,
-      ),
+      child: widget.isTodayWord == true
+          ? Text('오늘 학습한 단어 목록', style: TextStyles.xxLargeTextStyle,)
+          : Text('학습한 단어 목록', style: TextStyles.xxLargeTextStyle,),
     );
-
+    
     Widget wordListSection = FutureBuilder(
-        future: statement,
+        future: word,
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.waiting) {
             return Center(
@@ -134,63 +139,68 @@ class _TodayLearnStatementListPageState extends State<TodayLearnStatementListPag
                   )
               );
             } else {
-              List<Statement> statements = snapshot.data;
+              List<Word> words = snapshot.data;
               return Container(
                 height: MediaQuery.of(context).size.height*0.53,
-                margin: const EdgeInsets.symmetric(horizontal: 25),
+                margin: const EdgeInsets.symmetric(horizontal: 30),
                 child: ListView.separated(
                     itemBuilder: ((context, index) {
-                      Statement statement = statements[index];
-                      return ListTile(
-                        title: Container(
-                          padding: EdgeInsets.symmetric(vertical: 3),
-                          child: Text(
-                            statement.content,
+                      Word word = words[index];
+                      return Row(
+                        children: [
+                          Text(
+                            word.notation,
                             style: TextStyles.regular00TextStyle,
                           ),
-                        ),
-                        subtitle: Wrap(
-                          spacing: 5,
-                          children: statement.tags.map((tag) {
-                            return Chip(
-                                label: Text(tag),
-                                labelStyle: TextStyles.small00TextStyle,
-                                backgroundColor: selectTagColor(tag)
-                            );
-                          }).toList(),
-                        ),
-                        trailing: IconButton(
-                            onPressed: (){
-                              if(statement.bookmarked){
-                                setState(() {
-                                  statement.bookmarked = false;
-                                });
-                                deleteBookmark(statement.id);
-                              }
-                              else{
-                                setState(() {
-                                  statement.bookmarked = true;
-                                });
-                                createBookmark(statement.id);
-                              }
-                            },
-                            icon: statement.bookmarked?
-                            SvgPicture.asset(
-                              'assets/icons/star_fill.svg',
-                              fit: BoxFit.scaleDown,
-                            )
-                                :
-                            SvgPicture.asset(
-                              'assets/icons/star_unfill.svg',
-                              fit: BoxFit.scaleDown,
-                            )
-                        )
+                          Text(
+                            '[${word.pronunciation}]',
+                            style: TextStyles.regularGreenTextStyle,
+                          ),
+                          Spacer(flex: 2,),
+                          Chip(
+                            label: Text(
+                              word.tag,
+                              style: TextStyles.small00TextStyle,
+                            ),
+                            backgroundColor: selectTagColor(word.tag),
+                          ),
+                          Container(
+                            margin: EdgeInsets.only(left: MediaQuery.of(context).size.width*0.02),
+                            child: IconButton(
+                                onPressed: (){
+                                  if(word.bookmarked){
+                                    setState(() {
+                                      word.bookmarked = false;
+                                    });
+                                    deleteBookmark(word.id);
+                                  }
+                                  else{
+                                    setState(() {
+                                      word.bookmarked = true;
+                                    });
+                                    createBookmark(word.id);
+                                  }
+                                },
+                                icon: word.bookmarked?
+                                SvgPicture.asset(
+                                  'assets/icons/star_fill.svg',
+                                  fit: BoxFit.scaleDown,
+                                )
+                                    :
+                                SvgPicture.asset(
+                                  'assets/icons/star_unfill.svg',
+                                  fit: BoxFit.scaleDown,
+                                )
+                            ),
+                          )
+                        ],
+
                       );
                     }),
                     separatorBuilder: (BuildContext context, int index) {
                       return const Divider(thickness: 1, height: 3,);
                     },
-                    itemCount: widget.sentenceList.length
+                    itemCount: snapshot.data!.length
                 ),
               );
             }
@@ -200,41 +210,47 @@ class _TodayLearnStatementListPageState extends State<TodayLearnStatementListPag
 
     Widget retryLearnButtonSection = Container(
       margin: EdgeInsets.only(
-          top: MediaQuery.of(context).size.height*0.02,
-          left: MediaQuery.of(context).size.width*0.04,
-          right: MediaQuery.of(context).size.width*0.04
+        top: MediaQuery.of(context).size.height*0.02,
+        left: MediaQuery.of(context).size.width*0.04,
+        right: MediaQuery.of(context).size.width*0.04
       ),
       decoration: BoxDecoration(
-          boxShadow: [
-            BoxShadow(
-              color: Colors.grey.withOpacity(0.2),
-              spreadRadius: 6,
-              blurRadius: 7,
-              offset: Offset(0, 3),
-            )
-          ]
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.withOpacity(0.2),
+            spreadRadius: 6,
+            blurRadius: 7,
+            offset: Offset(0, 3),
+          )
+        ]
       ),
       child: OutlinedButton(
           onPressed: () {
-            Navigator.push(context, MaterialPageRoute(
-              builder: (context) => AccentTodayPracticePage(idx: 0, sentenceList: widget.sentenceList)
-            ));
+            if (widget.isTodayWord == true) {
+              Navigator.push(context, MaterialPageRoute(
+                builder: (context) => PronouncePracticePage(idx: 0, isTodayLearn: true, wordList: widget.wordList, pcList: [],), //이미 학습한것은 어떻게 처리? idx
+              ));
+            } else {
+              Navigator.push(context, MaterialPageRoute(
+                builder: (context) => PronouncePracticePage(idx: 0, isTodayLearn: false, wordList: widget.wordList, pcList: [],), //이미 학습한것은 어떻게 처리? idx
+              ));
+            }
           },
           style: OutlinedButton.styleFrom(
-              shape: const RoundedRectangleBorder(
-                  borderRadius: BorderRadius.all(Radius.circular(8.0))
-              ),
-              side: const BorderSide(
-                width: 1.0,
-                color: Colors.transparent,
-              ),
-              backgroundColor: Colors.white
+            shape: const RoundedRectangleBorder(
+                borderRadius: BorderRadius.all(Radius.circular(8.0))
+            ),
+            side: const BorderSide(
+              width: 1.0,
+              color: Colors.transparent,
+            ),
+            backgroundColor: Colors.white
           ),
           child: Container(
             padding: EdgeInsets.all(MediaQuery.of(context).size.height*0.023),
             child: const Text(
               '다시 학습',
-              style: TextStyles.mediumSTextStyle,
+              style: TextStyles.mediumRecordTextStyle,
               textAlign: TextAlign.center,
             ),
           )
@@ -243,22 +259,22 @@ class _TodayLearnStatementListPageState extends State<TodayLearnStatementListPag
 
     Widget goMainPageSection = Container(
       margin: EdgeInsets.only(
-          top: MediaQuery.of(context).size.height*0.01,
-          left: MediaQuery.of(context).size.width*0.04,
-          right: MediaQuery.of(context).size.width*0.04
+        top: MediaQuery.of(context).size.height*0.01,
+        left: MediaQuery.of(context).size.width*0.04,
+        right: MediaQuery.of(context).size.width*0.04
       ),
       child: OutlinedButton(
           onPressed: () => Navigator.pushAndRemoveUntil(context, MaterialPageRoute(
               builder: (BuildContext context) =>
                   HomePage()), (route) => false),
           style: OutlinedButton.styleFrom(
-            backgroundColor: ColorStyles.saeraRed,
+            backgroundColor: ColorStyles.saeraOlive1,
             shape: const RoundedRectangleBorder(
                 borderRadius: BorderRadius.all(Radius.circular(8.0))
             ),
             side: const BorderSide(
               width: 1.0,
-              color: ColorStyles.saeraRed,
+              color: ColorStyles.saeraOlive1,
             ),
           ),
           child: Container(
@@ -305,4 +321,13 @@ class _TodayLearnStatementListPageState extends State<TodayLearnStatementListPag
     );
   }
 
+}
+
+class Word {
+  final int id;
+  final String notation;
+  final String pronunciation;
+  final String tag;
+  bool bookmarked;
+  Word({required this.id, required this.notation, required this.pronunciation, required this.tag, required this.bookmarked});
 }
